gadget: update to a working boot.scr.in
Note that we comment out the loading of the device tree file. The reason
for this is because Qemu automatically passes its builtin device tree to
OpenSBI/u-boot at a specific address, and it would be useless to load an
identical dtb to the same address. So we won't bother. In the future,
these comments should be removed for real hardware.