gadget: snapcraft.yaml: update HSS
We don't actually need to build HSS. The HSS is flashed to the board
using FPExpress, a tool distributed by Microchip. The HSS we build is
never actually used, because the HSS flashed to the board using this
tool simply executes a payload binary. That payload binary is all we
really need to build.